requirements.txt
pytest
selene>=2.0.0rc7
allure-pytest